chore(LSPRefactored,LSPViolated, LSPViolatedTest): move package
