+++ app/src/test/java/com/example/learningtest/solid/lsp/LSPViolated.kt
@@ -1,4 +1,4 @@
-package com.example.learningtest.solid
+package com.example.learningtest.solid.lsp
 
 class LSPViolated {
     private class Customer {
